Modifications

+++ src/main/java/org/dcu/cloud/assignment/tfidf/TFIDFDriver.java
@@ -38,9 +38,9 @@ public class TFIDFDriver extends Configured implements Tool {
         tfidfCalculator.setMapperClass(TFIDFMapper.class);
         tfidfCalculator.setReducerClass(TFIDFWeightReducer.class);
         tfidfCalculator.setMapOutputKeyClass(Text.class);
-        tfidfCalculator.setMapOutputValueClass(DoubleWritable.class);
+        tfidfCalculator.setMapOutputValueClass(IntWritable.class);
         tfidfCalculator.setOutputKeyClass(Text.class);
-        tfidfCalculator.setOutputValueClass(DoubleWritable.class);
+        tfidfCalculator.setOutputValueClass(IntWritable.class);
         tfidfCalculator.setNumReduceTasks(1);
         FileInputFormat.addInputPath(tfidfCalculator, new Path(postWordCountPerDocOutPut));
         FileOutputFormat.setOutputPath(tfidfCalculator, new Path(perUserTFIDF));
